Make successive test runs work.
Since the tests were stateful, we have to clear out existing Factotum
commands between runs.  This is messy and it also breaks the extension once
the tests were run, since only the command(s) from the final run will be
available.  This implies that you shouldn't use the extension if you have
run the tests.  This should probably be cleaned up.