Refactor TransferObject to use a Proxy
- using a Proxy greatly simplifies the logic for handling direct property
  operations on a TransferObject.

However:  that simplification is not yet the only part of this commit.  In
order to more easily detect TransferObject usage that would need updating,
the proxy is much more complex, since it has to bridge between the previous
get/set/has/delete<method> calls and the actual properties to be operated
upon.

This uncovered a few bugs in transferobject data passing, now fixed.

Next up:  convert to using direct property access (since the proxy handler
will verify if the property is supported or not).